bits: fix gnome-terminal profile generator, Fix #35
Right now, a clean install (or a `dconf reset`) of the gnome-terminal
has a completely empty dconf configuration.

Adding the profile works, but the terminal won't recognize new profiles
if there isn't a `list` attribute with all the UUIDs of the profiles.

Also the `Unnamed Profile` has a fixed UUID
`b1dcc9dd-5262-4d8d-a863-c897e6d979b9`, so we must include that or it
won't work...

Signed-off-by: Luca Di Maio <luca.dimaio1@gmail.com>